refactor: Segregated into components (Vertical)
